backend/services/scheduler/requirements.txt
# FastAPI and dependencies
fastapi==0.119.0
uvicorn[standard]==0.34.1
python-dotenv==1.0.1
pydantic==2.12.1
pydantic-settings==2.7.1

# HTTP client
httpx==0.28.1

# Scheduler
apscheduler==3.10.4

# Utilities
python-multipart==0.0.20
